requirements.txt
# Core backend packages
fastapi==0.110.0
uvicorn==0.27.1
flask>=2.3.3
flask-cors>=4.0.0
python-dotenv>=1.0.0
Werkzeug>=2.3.7
google-generativeai>=0.3.1
llama-index-core>=0.10.0
llama_index
# llama-index-llms-vertex
watchdog>=3.0.0
pydantic>=2.8.0,<3.0.0

# Google Cloud dependencies
google-cloud-storage>=2.13.0
google-cloud-logging>=3.8.0
google-cloud-bigquery>=3.19.0
google-auth-oauthlib>=1.2.2
google-api-python-client>=2.0.0
# Remove the fixed version to allow pip to resolve dependencies
google-cloud-aiplatform>=1.71.0
# vertexai>=0.1.0
# llama-index-embeddings-vertex>=0.1.0
llama-index-embeddings-google-genai>=0.1.0
llama-index-llms-google-genai>=0.1.0
llama-index-readers-docling>=0.1.0
dotenv>=0.9.9
concurrently
pandas-gbq
db-dtypes